jj_no_fish/requirements.txt
# Web Framework
Django==5.0.1

# Document Processing
python-docx==1.0.1

# Static Files Management
whitenoise==6.6.0
django-compressor==4.4    # For CSS/JS compression
django-pipeline==2.1.0    # For asset packaging

# Development Tools
django-debug-toolbar==4.2.0
django-livereload-server==0.5.1  # For auto-reloading during development

# Testing Dependencies
pytest==7.4.4
pytest-django==4.7.0

# Security
django-cors-headers==4.3.1

# Python Utilities
python-dotenv==1.0.0